fixed and refactored savings screen
