+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/screens/intro_navbar/intro_products/SavingsScreen.kt
@@ -1,4 +1,4 @@
-package com.semirsuljevic.raiffaisenmobileapp.ui.screens
+package com.semirsuljevic.raiffaisenmobileapp.ui.screens.intro_navbar.intro_products
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -18,6 +19,7 @@ import androidx.navigation.NavController
 import com.semirsuljevic.raiffaisenmobileapp.R
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.CenteredTitleAppBar
 import com.semirsuljevic.raiffaisenmobileapp.ui.composables.ListItemSeparator
+import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Black
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Gray200
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Yellow400
 
@@ -26,24 +28,33 @@ fun SavingsScreen(navController: NavController) {
     val scrollState = rememberScrollState()
     val uriHandler = LocalUriHandler.current
     val moreInfoLink = stringResource(id = R.string.savings_screen_more_info_link)
-    Column (
-        Modifier.verticalScroll(scrollState)
-    ){
-        CenteredTitleAppBar(title = stringResource(id = R.string.savings_screen_title), navController = navController)
-        ListItemSeparator()
-        Spacer(modifier = Modifier.height(20.dp))
-        Column (Modifier.padding(horizontal = 20.dp)){
-            Text(
-                text = stringResource(id = R.string.savings_screen_description),
-                color = Gray200,
-                textAlign = TextAlign.Justify
+    Scaffold (
+        backgroundColor = Black,
+        topBar = {
+            CenteredTitleAppBar(
+                title = stringResource(id = R.string.savings_screen_title),
+                navController = navController
             )
+        }
+    ) {
+        Column (
+            Modifier.verticalScroll(scrollState)
+        ){
+            ListItemSeparator()
             Spacer(modifier = Modifier.height(20.dp))
-            Text(
-                text = stringResource(id = R.string.savings_screen_more_info),
-                color = Yellow400,
-                modifier = Modifier.clickable { uriHandler.openUri(moreInfoLink) }
-            )
+            Column (Modifier.padding(horizontal = 20.dp)){
+                Text(
+                    text = stringResource(id = R.string.savings_screen_description),
+                    color = Gray200,
+                    textAlign = TextAlign.Justify
+                )
+                Spacer(modifier = Modifier.height(20.dp))
+                Text(
+                    text = stringResource(id = R.string.savings_screen_more_info),
+                    color = Yellow400,
+                    modifier = Modifier.clickable { uriHandler.openUri(moreInfoLink) }
+                )
+            }
         }
     }
 }
